refactored product class, controller added ProductDTO and UserService for user retrieval

+++ auction-app/src/main/java/com/auctionapp/controller/ProductController.java
@@ -1,10 +1,15 @@
 package com.auctionapp.controller;
 
+import com.auctionapp.model.product.EProductCategory;
 import com.auctionapp.model.product.Product;
+import com.auctionapp.model.product.ProductDTO;
 import com.auctionapp.service.ProductService;
+import com.auctionapp.service.UserService;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -14,45 +19,94 @@ public class ProductController {
 
     public final ProductService productService;
 
-    public ProductController(ProductService productService) {
+    public final UserService userService;
+
+    public ProductController(ProductService productService, UserService userService) {
         this.productService = productService;
+        this.userService = userService;
     }
 
-    @GetMapping()
-    public ResponseEntity<List<Product>> getAllProducts() {
+    @GetMapping("/")
+    public ResponseEntity<List<ProductDTO>> getAllProducts() {
         List<Product> products = productService.getAllProducts();
-        return ResponseEntity.ok(products);
+        List<ProductDTO> listOfProductResponses = convertToDTOList(products);
+        return ResponseEntity.ok(listOfProductResponses);
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Product> getProductById(@PathVariable Long id) {
-        Optional<Product> optionalProduct = productService.getProductById(id);
-        return optionalProduct.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
+    public ResponseEntity<ProductDTO> getProductById(@PathVariable Long id) {
+        Optional<Product> oProduct = productService.getProductById(id);
+        return oProduct.map(product -> ResponseEntity.ok(convertToDTO(product)))
+                .orElse(ResponseEntity.notFound().build());
     }
 
     @PostMapping("/add")
-    public ResponseEntity<Product> createProduct(@RequestBody Product product) {
+    public ResponseEntity<String> createProduct(@RequestBody Product product) {
         if (product.getId() != null && product.getId() != 0) {
             return ResponseEntity.badRequest().build();
         }
-
-        Product createdProduct = productService.createProduct(product);
-        return ResponseEntity.ok(createdProduct);
+        Long userId = product.getUser().getId();
+        if (userService.getUserById(userId).isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User for adding product doesn't exist");
+        }
+        productService.createProduct(product);
+        return ResponseEntity.ok("Product created successfully!");
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Product> updateProduct(@PathVariable Long id, @RequestBody Product product) {
-        if (product.getId() == null || product.getId() == 0) {
-            return ResponseEntity.badRequest().build();
+    public ResponseEntity<String> updateProduct(@PathVariable Long id, @RequestBody ProductDTO productDTO) {
+        if (id == null || productDTO.getUserId() == null) {
+            return ResponseEntity.badRequest().body("Invalid request data!");
         }
-        Product updatedProduct = productService.updateProduct(product);
-        return ResponseEntity.ok(updatedProduct);
+
+        var oExistingProduct = productService.getProductById(id);
+        if (oExistingProduct.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
+        var existingProduct = oExistingProduct.get();
+        existingProduct.setName(productDTO.getName());
+        existingProduct.setStarting_price(productDTO.getStarting_price());
+        existingProduct.setDescription(productDTO.getDescription());
+        existingProduct.setCategory(EProductCategory.valueOf(productDTO.getCategory()));
+        Long userId = productDTO.getUserId();
+        if (userService.getUserById(userId).isPresent()) {
+            existingProduct.getUser().setId(productDTO.getUserId());
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User for updating product doesn't exists!");
+        }
+        productService.updateProduct(existingProduct);
+        return ResponseEntity.ok("Product was updated successfully!");
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<String> deleteProduct(@PathVariable Long id) {
-        productService.deleteProduct(id);
-        return ResponseEntity.ok("Product was deleted!");
+        if (id != null && productService.getProductById(id).isPresent()) {
+            productService.deleteProduct(id);
+            return ResponseEntity.ok("Product was deleted successfully!");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Product for deletion doesn't exists!");
+        }
+    }
+
+    private ProductDTO convertToDTO(Product product) {
+        ProductDTO productDTO = new ProductDTO();
+        productDTO.setId(product.getId());
+        productDTO.setName(product.getName());
+        productDTO.setStarting_price(product.getStarting_price());
+        productDTO.setDescription(product.getDescription());
+        productDTO.setCategory(String.valueOf(product.getCategory()));
+        if (product.getUser() != null) {
+            productDTO.setUserId(product.getUser().getId());
+        }
+        return productDTO;
+    }
+
+    private List<ProductDTO> convertToDTOList(List<Product> products) {
+        List<ProductDTO> productDTOList = new ArrayList<>();
+        for (Product product : products) {
+            productDTOList.add(convertToDTO(product));
+        }
+        return productDTOList;
     }
 
 }
